Touch Event
Implemented touch event under -(void)touchesBegan - this is called
implicit animation because it the changes to some properties are
automatically animated when the setters are called - position is one of
these implicitly animatable properties. Moving on to a tap event - I
think I'll have the tap on the subview "sandboxView" bring us to a
camera chooser.