Change Super on Click
Cool tweak - everytime the touch begins it changes the super view.
I solidified my knowledge on using the @property and @synthesize
directives (?)

Now I noticed something. If I want to change the text now I can't
because (I think) the first responder is now the touch event. I may
have to get the touch coordinates and if it is on top of the button
then execute the button logic...I wonder if that's going to be the
solution.

Here's to hoping it's that easy...